# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NanoPay API
  version: "2.0.0"
  description: |
    API Backend NanoPay — commande centrale unique reliant Client, Commerçant, Banque et Admin,
    avec trois modes de financement : **Épargne**, **Coffre** (prélèvement sur salaire) et **Crédit**.

    Conforme au « Cahier de construction BACKEND ».

    ## Contrat de connexion
    | Élément | Contrat |
    |---|---|
    | Base URL | `/api` (alias historique `/api/v1`) |
    | Format | JSON UTF-8 (multipart pour les documents) |
    | Auth | `Authorization: Bearer <accessToken>` + refresh token (`/auth/refresh`, rotation) |
    | Succès | `{ "data": ..., "meta"?: ... }` |
    | Erreur | `{ "code", "message", "details"?, "requestId" }` (`X-Request-Id` aussi en en-tête) |
    | Pagination | query `page`, `limit` ; `meta` = `{ page, limit, total, nextPage }` |
    | Dates | ISO 8601 |
    | Montants | FCFA **entiers**, jamais de flottant |
    | Identifiants | UUID serveur ; `orderNumber` lisible (ex. `NP-2026-00001245`) pour l'UI |
    | Temps réel | pas de WebSocket : polling contrôlé de `GET /notifications` et des ressources |

    Les règles financières et de sécurité sont **toujours** appliquées côté backend ; le frontend
    ne fait que consommer ces contrats.

    ## Circuit
    Client → produit → commande (`CREEE`) → `select-financing` (`FINANCEMENT_EN_COURS`) →
    Épargne / Coffre / Crédit → `FINANCEE` → `PRETE_A_LIVRER` (+ QR) → scan commerçant →
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /credit/profile:
    get:
      tags: [Credit]
      summary: Mon profil crédit (`NON_CONFIGURE` s'il n'existe pas)
      responses:
        "200": { description: Profil }
    post:
      tags: [Credit]
      summary: Soumet le profil crédit à la banque
      description: "`multipart` : `employer`, `registrationNumber?`, `bankAuthorization` (fichier), `paySlips` (fichier). Statut → `EN_ANALYSE`."
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [employer, bankAuthorization, paySlips]
              properties:
                employer: { type: string }
                registrationNumber: { type: string }
                bankAuthorization: { type: string, format: binary }
                paySlips: { type: string, format: binary }
      responses:
        "201": { description: Profil soumis }
        "409": { $ref: "#/components/responses/Error" }
  /credit/requests:
    post:
      tags: [Credit]
      summary: Crée la demande de crédit d'une commande
      description: "Profil `VALIDE` requis. Durée validée par le backend (≤50 000 : 8 mois ; 50 000–100 000 : 12 ; au-delà +6 mois par tranche de 50 000). Montant à virer = prix + frais."
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [orderId, durationMonths], properties: { orderId: { type: string }, durationMonths: { type: integer } } } } }
      responses:
        "201": { description: Demande `ENVOYEE` }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /credit/requests/{id}:
    get:
      tags: [Credit]
      summary: Détail d'une demande de crédit
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Demande }
  /credits/{id}:
    get:
      tags: [Credit]
      summary: Détail d'un crédit
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Crédit }
  # ------------------------------------------------------------------ MERCHANT
...
  /bank/credit-profiles:
    get:
      tags: [Bank]
      summary: Profils crédit assignés (par défaut `EN_ANALYSE`)
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [NON_CONFIGURE, EN_ANALYSE, VALIDE, REFUSE] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page de profils } }
  /bank/credit-profiles/{id}/decision:
    patch:
      tags: [Bank]
      summary: Valide ou refuse un profil
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [decision], properties: { decision: { type: string, enum: [VALIDE, REFUSE] }, reason: { type: string } } } } }
      responses: { "200": { description: Décision enregistrée }, "409": { $ref: "#/components/responses/Error" } }
  /bank/credit-requests:
    get:
      tags: [Bank]
      summary: Demandes de crédit assignées (par défaut `ENVOYEE` + `EN_ANALYSE`)
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [BROUILLON, ENVOYEE, EN_ANALYSE, ACCEPTEE, REFUSEE] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page de demandes } }
  /bank/credit-requests/{id}/decision:
    patch:
      tags: [Bank]
      summary: Accepte ou refuse une demande
      description: "`ACCEPTEE` crée le Crédit `EN_ATTENTE_DE_VIREMENT`. `REFUSEE` remet la commande à `CREEE`."
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [decision], properties: { decision: { type: string, enum: [ACCEPTEE, REFUSEE] }, reason: { type: string } } } } }
      responses: { "200": { description: Décision enregistrée }, "409": { $ref: "#/components/responses/Error" } }
  /bank/transfers/notice:
    post:
      tags: [Bank]
      summary: Avis de virement banque → NanoPay
      description: |
        `amount` doit égaler exactement le montant à virer (prix + frais), sinon 422. `transferReference` est UNIQUE :
        un même virement n'est jamais comptabilisé deux fois. Crédit `FINANCE`, commande `FINANCEE` → `PRETE_A_LIVRER` + QR.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [creditId, transferReference, amount]
              properties:
                creditId: { type: string }
                transferReference: { type: string }
                amount: { type: integer }
      responses: { "201": { description: Virement enregistré }, "422": { $ref: "#/components/responses/Error" } }
  # ------------------------------------------------------------------ ADMIN
...
  /admin/clients:
    get:
      tags: [Admin]
      summary: Clients
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [ACTIF, SUSPENDU] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page de clients } }
  /admin/clients/{id}/status:
    patch:
      tags: [Admin]
      summary: Suspend / réactive un client
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [ACTIF, SUSPENDU] }, reason: { type: string } } } } }
      responses: { "200": { description: Client mis à jour } }
  /admin/merchants:
    get:
      tags: [Admin]
      summary: Commerçants
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [EN_ATTENTE, VALIDE, SUSPENDU] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page de commerçants } }
  /admin/merchants/{id}/status:
    patch:
      tags: [Admin]
      summary: Valide / suspend un commerçant
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [status], properties: { status: { type: string, enum: [VALIDE, SUSPENDU] }, reason: { type: string } } } } }
      responses: { "200": { description: Commerçant mis à jour } }
  /admin/savings:
    get:
      tags: [Admin]
      summary: Épargnes
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses: { "200": { description: Page d'épargnes } }
  /admin/savings/{id}/refund:
    post:
      tags: [Admin]
      summary: Marque le remboursement (85 %) d'une épargne `ECHOUEE` comme versé → `REMBOURSEE`
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { reference: { type: string } } } } }
      responses: { "200": { description: Remboursée }, "409": { $ref: "#/components/responses/Error" } }
  /admin/credits:
    get:
      tags: [Admin]
      summary: Crédits
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses: { "200": { description: Page de crédits } }
  /admin/payments:
    get:
      tags: [Admin]
      summary: Paiements
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [INITIE, EN_ATTENTE, CONFIRME, ECHOUE] } }
        - { name: orderId, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page de paiements } }
  /admin/audit-logs:
    get:
      tags: [Admin]
      summary: Journal d'audit (date, heure, identifiant, acteur, montant, statut)
      parameters:
        - { name: entityType, in: query, schema: { type: string } }
        - { name: entityId, in: query, schema: { type: string } }
        - { name: action, in: query, schema: { type: string } }
        - { name: userId, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page d'entrées d'audit } }
  /admin/settlements:
    get:
      tags: [Admin]
      summary: Règlements commerçants
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [EN_ATTENTE, PAYE] } }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/Limit" }
      responses: { "200": { description: Page de règlements } }
  /admin/settlements/{id}/paid:
    post:
      tags: [Admin]
      summary: Confirme le versement du règlement → commande `TERMINEE`, crédit `TERMINE`
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [reference], properties: { reference: { type: string } } } } }
      responses: { "200": { description: Réglé }, "409": { $ref: "#/components/responses/Error" } }
  /admin/settings/financial:
    get: { tags: [Admin], summary: Paramètres financiers effectifs, responses: { "200": { description: Paramètres } } }
    patch:
      tags: [Admin]
      summary: Modifie les paramètres financiers (pénalité, commissions, paliers, durées, TTL du QR…)
      description: Corps partiel ; toute clé inconnue est refusée (422). Tracé dans l'audit.
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/FinancialParams" } } }
      responses: { "200": { description: Paramètres mis à jour } }
  # ------------------------------------------------------------------ WEBHOOKS
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Error:
      description: "Erreur — `{ code, message, details?, requestId }`"
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        details: {}
        requestId: { type: string }
...
    FinancialParams:
      type: object
      properties:
        savingsTiers:
          type: array
          items:
            type: object
            properties:
              maxAmount: { type: integer, nullable: true }
              minInstallment: { type: integer }
              maxDurationMonths: { type: integer }
        savingsExtensionMaxMonths: { type: integer }
        savingsPenaltyRate: { type: number, example: 0.15 }
        savingsCommissionRates:
          type: object
          properties: { tier1: { type: number }, tier2: { type: number }, tier3: { type: number } }
        creditTier1MaxMonths: { type: integer }
        creditTier2MaxMonths: { type: integer }
        creditExtraMonthsPerBracket: { type: integer }
        creditBracketSize: { type: integer }
        creditBankFeeRate: { type: number, example: 0.02 }
        creditMerchantCommissionRate: { type: number, example: 0.04 }
        qrTtlHours: { type: integer, example: 72 }